Extract widget presentation module
